fix(Board): 점수 계산 기능 내의 지역 변수 타입 변경 ( int -> double )

+++ src/main/java/softeer2nd/chess/Board.java
@@ -149,7 +149,7 @@ public class Board {
     }
 
     public double calculatePoint(PieceColor color) {
-        int point = 0;
+        double point = 0;
         for (List<Piece> rank : boards) {
             int numOfPawns = 0;
             for (Piece piece : rank) {
